+++ community/components/navigation/vertical-stepper/index.ts
@@ -1,0 +1,2 @@
+export * from "./vertical-stepper.component";
+export * from "./vertical-stepper-item/vertical-stepper-item.component";