+++ tedi/services/index.ts
@@ -2,3 +2,4 @@ export * from "./breakpoint/breakpoint.service";
 export * from "./translation/translation.service";
 export * from "./theme/theme.service";
 export * from "./translation/translation.pipe";
+export * from "./toast/toast.service";